requirements.txt
astroid==1.6.4
autopep8==1.3.5
certifi==2018.4.16
chardet==3.0.4
cloverapi-python==0.0.1
Django==1.11.3
django-crispy-forms==1.7.2
django-multiselectfield==0.1.8
django-resized==0.3.9
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.1.0
pluggy==0.7.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.4
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==1.9.2
PySocks==1.6.8
python-resize-image==1.1.11
pytz==2018.4
requests==2.13.0
six==1.11.0
sparkpost==1.3.6
tox==3.2.1
twilio==6.16.1
urllib3==1.23
virtualenv==16.0.0
wrapt==1.10.11
